Split client tests into unit tests and end-to-end tests
The client tests are very slow to run and have external dependencies on a
running dataserv server.

This change splits up the client tests into end-to-end tests and unit tests,
where unit tests have no dependencies on external processes or global state.

The end-to-end tests exercise client-server interaction, but do not test
scenarios that are client or server focused. For example, there is no end to
end tests for double-registration of the same address because this is covered
by unit tests in both dataserv-client and dataserv (server). The end to end
tests are mainly positive tests of successful interaction between client and
server.